requirements.txt
image==1.3.4
Pillow==2.7.0
pyserial==2.7
